if the private key file not exist , create it .

+++ bar_tunnel/common/generate_keys.py
@@ -22,42 +22,37 @@ def change_directory(ex_file,back,file):
     path = os.path.abspath(os.path.join(dir_of_executable, back)) + file
     return path
 
+def change_directory1(ex_file,back):
+    dir_of_executable = os.path.dirname(ex_file)
+    path = os.path.abspath(os.path.join(dir_of_executable, back))
+    return path
+
+
 def generate_rsa_key(bit):
     path_to_private_key = change_directory(__file__,"../../keys","/private_key.pem")
 
     key = RSA.generate(bit)
-    try:
-        f = open(path_to_private_key, 'r')
-    except IOError:
+    if os.path.exist(change_directory1(__file__,"../../keys")):
         f=  open(path_to_private_key,'w')
+    else:
+        os.makedirs(change_directory1(__file__,"../../keys"))
+        f = open(path_to_private_key, 'w')
     f.write(key.exportKey('PEM'))
     f.close()
 
     path_to_public_key = change_directory(__file__,"../../keys","/public_key.pem")
 
-    try:
-        f = open(path_to_public_key, 'r')
-    except IOError:
-        f=  open(path_to_public_key,'w')
-
+    f=  open(path_to_public_key,'w')
     f.write(key.publickey().exportKey())
     f.close()
 
 def generate_rsa_bridge_key(bit):
     key = RSA.generate(bit)
-    try:
-        f = open('keys/bridge_private_key.pem', 'r')
-    except IOError:
-        f = open('keys/bridge_private_key.pem', 'w')
-
+    f=  open('keys/bridge_private_key.pem','w')
     f.write(key.exportKey('PEM'))
     f.close()
 
-    try:
-        f = open('keys/bridge_public_key.pem', 'r')
-    except IOError:
-        f = open('keys/bridge_public_key.pem', 'w')
-
+    f=  open('keys/bridge_public_key.pem','w')
     f.write(key.publickey().exportKey())
     f.close()
